Need to dispatchComponentURL() within router.ts (in AppNavigation it is too late)
