+++ aas-web-ui/src/components/AppNavigation/AppNavigation.vue
@@ -193,8 +193,7 @@
 </template>
 
 <script lang="ts" setup>
-    import type { BaSyxComponent, RepositoryKey } from '@/types/BaSyx';
-    import { computed, mergeProps, onMounted, reactive, ref, watch } from 'vue';
+    import { computed, mergeProps, onMounted, ref, watch } from 'vue';
     import { useRoute } from 'vue-router';
     import { useTheme } from 'vuetify';
     import { useDashboardHandling } from '@/composables/DashboardHandling';
@@ -216,51 +215,48 @@
     // Vuetify
     const theme = useTheme();
 
-    // Reactive BaSyx Components Configurations
-    const basyxComponents = reactive<Record<RepositoryKey, BaSyxComponent>>({
-        AASDiscovery: {
-            url: ref(navigationStore.getAASDiscoveryURL), // Ensure the getter is invoked
-            loading: ref(false),
-            connect: () => connectComponent('AASDiscovery'),
-            label: 'AAS Discovery URL',
-        },
-        AASRegistry: {
-            url: ref(navigationStore.getAASRegistryURL),
-            loading: ref(false),
-            connect: () => connectComponent('AASRegistry'),
-            label: 'AAS Registry URL',
-        },
-        SubmodelRegistry: {
-            url: ref(navigationStore.getSubmodelRegistryURL),
-            loading: ref(false),
-            connect: () => connectComponent('SubmodelRegistry'),
-            label: 'Submodel Registry URL',
-        },
-        AASRepo: {
-            url: ref(navigationStore.getAASRepoURL),
-            loading: ref(false),
-            connect: () => connectComponent('AASRepo'),
-            label: 'AAS Repository URL',
-        },
-        SubmodelRepo: {
-            url: ref(navigationStore.getSubmodelRepoURL),
-            loading: ref(false),
-            connect: () => connectComponent('SubmodelRepo'),
-            label: 'Submodel Repository URL',
-        },
-        ConceptDescriptionRepo: {
-            url: ref(navigationStore.getConceptDescriptionRepoURL),
-            loading: ref(false),
-            connect: () => connectComponent('ConceptDescriptionRepo'),
-            label: 'Concept Description Repository URL',
-        },
-    });
+    //         url: ref(navigationStore.getAASDiscoveryURL), // Ensure the getter is invoked
+    //         loading: ref(false),
+    //         connect: () => connectComponent('AASDiscovery'),
+    //         label: 'AAS Discovery URL',
+    //     },
+    //     AASRegistry: {
+    //         url: ref(navigationStore.getAASRegistryURL),
+    //         loading: ref(false),
+    //         connect: () => connectComponent('AASRegistry'),
+    //         label: 'AAS Registry URL',
+    //     },
+    //     SubmodelRegistry: {
+    //         url: ref(navigationStore.getSubmodelRegistryURL),
+    //         loading: ref(false),
+    //         connect: () => connectComponent('SubmodelRegistry'),
+    //         label: 'Submodel Registry URL',
+    //     },
+    //     AASRepo: {
+    //         url: ref(navigationStore.getAASRepoURL),
+    //         loading: ref(false),
+    //         connect: () => connectComponent('AASRepo'),
+    //         label: 'AAS Repository URL',
+    //     },
+    //     SubmodelRepo: {
+    //         url: ref(navigationStore.getSubmodelRepoURL),
+    //         loading: ref(false),
+    //         connect: () => connectComponent('SubmodelRepo'),
+    //         label: 'Submodel Repository URL',
+    //     },
+    //     ConceptDescriptionRepo: {
+    //         url: ref(navigationStore.getConceptDescriptionRepoURL),
+    //         loading: ref(false),
+    //         connect: () => connectComponent('ConceptDescriptionRepo'),
+    //         label: 'Concept Description Repository URL',
+    //     },
+    // });
 
     // Data
     const mainMenu = ref(false); // Variable to show the Main Menu
     const mobileMenu = ref(false); // Variable to show the Mobile Menu
     const dashboardAvailable = ref(false); // Dashboard Availability
-    const endpointConfigAvailable = ref(envStore.getEndpointConfigAvailable);
+    // const endpointConfigAvailable = ref(envStore.getEndpointConfigAvailable);
     const drawerVisibility = ref(true); // Variable to show the AAS List Drawer
 
     // Computed Properties
@@ -269,12 +265,12 @@
     const Snackbar = computed(() => navigationStore.getSnackbar);
     const showAASList = computed(() => ['AASViewer', 'AASEditor', 'SubmodelViewer'].includes(route.name as string));
     const drawerState = computed(() => navigationStore.getDrawerState);
-    const EnvAASDiscoveryPath = computed(() => envStore.getEnvAASDiscoveryPath);
-    const EnvAASRegistryPath = computed(() => envStore.getEnvAASRegistryPath);
-    const EnvSubmodelRegistryPath = computed(() => envStore.getEnvSubmodelRegistryPath);
-    const EnvAASRepoPath = computed(() => envStore.getEnvAASRepoPath);
-    const EnvSubmodelRepoPath = computed(() => envStore.getEnvSubmodelRepoPath);
-    const EnvConceptDescriptionRepoPath = computed(() => envStore.getEnvConceptDescriptionRepoPath);
+    // const EnvAASDiscoveryPath = computed(() => envStore.getEnvAASDiscoveryPath);
+    // const EnvAASRegistryPath = computed(() => envStore.getEnvAASRegistryPath);
+    // const EnvSubmodelRegistryPath = computed(() => envStore.getEnvSubmodelRegistryPath);
+    // const EnvAASRepoPath = computed(() => envStore.getEnvAASRepoPath);
+    // const EnvSubmodelRepoPath = computed(() => envStore.getEnvSubmodelRepoPath);
+    // const EnvConceptDescriptionRepoPath = computed(() => envStore.getEnvConceptDescriptionRepoPath);
     const LogoPath = computed(() => {
         if (isDark.value && envStore.getEnvLogoDarkPath.trim().length > 0) {
             return 'Logo/' + envStore.getEnvLogoDarkPath;
@@ -318,48 +314,48 @@
         dashboardAvailable.value = await checkDashboardAvailability();
         applyTheme();
 
-        // Auto connect to BaSyx Components
-        Object.keys(basyxComponents).forEach((key) => {
-            const repoKey = key as RepositoryKey;
-            const storedURL = window.localStorage.getItem(repoKey);
+        // // Auto connect to BaSyx Components
+        // Object.keys(basyxComponents).forEach((key) => {
+        //     const repoKey = key as RepositoryKey;
+        //     const storedURL = window.localStorage.getItem(repoKey);
 
-            // console.log('storedURL: ', storedURL, repoKey);
+        //     // console.log('storedURL: ', storedURL, repoKey);
 
-            if (endpointConfigAvailable.value && storedURL) {
-                basyxComponents[repoKey].url = storedURL;
-                basyxComponents[repoKey].connect();
-            } else {
-                // Check environment path
-                let envPath = '';
-                switch (repoKey) {
-                    case 'AASDiscovery':
-                        envPath = EnvAASDiscoveryPath.value;
-                        break;
-                    case 'AASRegistry':
-                        envPath = EnvAASRegistryPath.value;
-                        break;
-                    case 'SubmodelRegistry':
-                        envPath = EnvSubmodelRegistryPath.value;
-                        break;
-                    case 'AASRepo':
-                        envPath = EnvAASRepoPath.value;
-                        break;
-                    case 'SubmodelRepo':
-                        envPath = EnvSubmodelRepoPath.value;
-                        break;
-                    case 'ConceptDescriptionRepo':
-                        envPath = EnvConceptDescriptionRepoPath.value;
-                        break;
-                    default:
-                        break;
-                }
+        //     if (endpointConfigAvailable.value && storedURL) {
+        //         basyxComponents[repoKey].url = storedURL;
+        //         basyxComponents[repoKey].connect();
+        //     } else {
+        //         // Check environment path
+        //         let envPath = '';
+        //         switch (repoKey) {
+        //             case 'AASDiscovery':
+        //                 envPath = EnvAASDiscoveryPath.value;
+        //                 break;
+        //             case 'AASRegistry':
+        //                 envPath = EnvAASRegistryPath.value;
+        //                 break;
+        //             case 'SubmodelRegistry':
+        //                 envPath = EnvSubmodelRegistryPath.value;
+        //                 break;
+        //             case 'AASRepo':
+        //                 envPath = EnvAASRepoPath.value;
+        //                 break;
+        //             case 'SubmodelRepo':
+        //                 envPath = EnvSubmodelRepoPath.value;
+        //                 break;
+        //             case 'ConceptDescriptionRepo':
+        //                 envPath = EnvConceptDescriptionRepoPath.value;
+        //                 break;
+        //             default:
+        //                 break;
+        //         }
 
-                if (!basyxComponents[repoKey].url && envPath && envPath.trim() !== '') {
-                    basyxComponents[repoKey].url = envPath;
-                    basyxComponents[repoKey].connect();
-                }
-            }
-        });
+        //         if (!basyxComponents[repoKey].url && envPath && envPath.trim() !== '') {
+        //             basyxComponents[repoKey].url = envPath;
+        //             basyxComponents[repoKey].connect();
+        //         }
+        //     }
+        // });
     });
 
     function closeSnackbar() {
@@ -390,9 +386,9 @@
         }
     }
 
-    function connectComponent(repoKey: keyof typeof basyxComponents) {
-        navigationStore.dispatchComponentURL(repoKey, basyxComponents[repoKey].url, false);
-    }
+    // function connectComponent(repoKey: keyof typeof basyxComponents) {
+    //     navigationStore.dispatchComponentURL(repoKey, basyxComponents[repoKey].url, false);
+    // }
 
     function extendSidebar() {
         drawerVisibility.value = true;
